12199: Add node size calculator.
Arvados-DCO-1.1-Signed-off-by: Tom Clegg <tclegg@veritasgenetics.com>

+++ sdk/go/arvados/container.go
@@ -41,9 +41,9 @@ type Mount struct {
 // CPU) and network connectivity.
 type RuntimeConstraints struct {
 	API          *bool
-	RAM          int `json:"ram"`
-	VCPUs        int `json:"vcpus"`
-	KeepCacheRAM int `json:"keep_cache_ram"`
+	RAM          int64 `json:"ram"`
+	VCPUs        int   `json:"vcpus"`
+	KeepCacheRAM int64 `json:"keep_cache_ram"`
 }
 
 // SchedulingParameters specify a container's scheduling parameters
